Adjust SQLite `qso` schema and model to use `types.Decimal` for `freq`
- Updated `freq` column type in SQLite `qso` schema and migrations to `DECIMAL`.
- Replaced `float64` with `types.Decimal` in the corresponding Go model.
- Updated helper functions to accommodate the new type.
- Refined `Taskfile.yml` to include Docker cleanup and clean tasks.
- Minor README adjustment: clarified directory context for SQLBoiler.

+++ sqlite/migrations/0001_schema_migrations.up.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS qso
     call            VARCHAR(20) NOT NULL,
     band            VARCHAR(10) NOT NULL,
     mode            VARCHAR(10) NOT NULL,
-    freq            REAL        NOT NULL CHECK (freq >= 0 AND freq <= 999.999 AND round(freq * 1000) = freq * 1000),
+    freq            DECIMAL NOT NULL CHECK (freq >= 0 AND freq <= 999.999 AND round(freq * 1000) = freq * 1000),
     /* DATETIME here prompts SQLBoiler to use the "time.Time" type */
     qso_date        DATETIME    NOT NULL CHECK (
         length(qso_date) = 8 AND
